:recycle: Rename INDEXER -> ENDPOINT, move utils
