+++ backend/package.json
@@ -14,6 +14,7 @@
     "@discordjs/builders": "^1.1.0",
     "@discordjs/rest": "^1.0.1",
     "@likecoin-discord-bot/config": "1.0.0",
+    "axios": "^0.27.2",
     "bcrypt": "^5.0.1",
     "bech32": "^2.0.0",
     "discord-api-types": "^0.37.1",
