+++ frontend/config.js
@@ -1,1 +1,2 @@
 export * from 'likecoin-discord-bot/config'
+export const ENDPOINT = ''
